Endringer etter pr-reviewet
