+++ src/main/kotlin/no/nav/familie/ef/sak/service/VurderingService.kt
@@ -7,6 +7,7 @@ import no.nav.familie.ef.sak.regler.HovedregelMetadata
 import no.nav.familie.ef.sak.regler.Vilkårsregel
 import no.nav.familie.ef.sak.regler.alleVilkårsregler
 import no.nav.familie.ef.sak.regler.evalutation.OppdaterVilkår
+import no.nav.familie.ef.sak.regler.evalutation.OppdaterVilkår.opprettNyeVilkårsvurderinger
 import no.nav.familie.ef.sak.regler.hentVilkårsregel
 import no.nav.familie.ef.sak.repository.VilkårsvurderingRepository
 import no.nav.familie.ef.sak.repository.domain.*
@@ -58,7 +59,7 @@ class VurderingService(private val behandlingService: BehandlingService,
 
 
     @Transactional
-    fun ikkeVurderVilkår(vilkårsvurderingDto: OppdaterVilkårsvurderingDto): VilkårsvurderingDto {
+    fun settVilkårTilSkalIkkeVurderes(vilkårsvurderingDto: OppdaterVilkårsvurderingDto): VilkårsvurderingDto {
         val vilkårsvurdering = vilkårsvurderingRepository.findByIdOrThrow(vilkårsvurderingDto.id)
         val behandlingId = vilkårsvurdering.behandlingId
 
@@ -67,7 +68,7 @@ class VurderingService(private val behandlingService: BehandlingService,
 
         blankettRepository.deleteById(behandlingId)
 
-        val oppdatertVilkår = oppdaterVilkårsvurderingTilIkkeVurder(behandlingId, vilkårsvurdering)
+        val oppdatertVilkår = oppdaterVilkårsvurderingTilSkalIkkeVurderes(behandlingId, vilkårsvurdering)
         oppdaterStegPåBehandling(behandlingId)
         return oppdatertVilkår
     }
@@ -85,12 +86,12 @@ class VurderingService(private val behandlingService: BehandlingService,
                 .filter { it.type != VilkårType.TIDLIGERE_VEDTAKSPERIODER } // TODO: Må håndteres senere
         val vilkårstyper = VilkårType.hentVilkår().minus(VilkårType.TIDLIGERE_VEDTAKSPERIODER)
         val vilkårUtenVurdering =
-                filtereVilkårMedResultat(behandlingId,
+                OppdaterVilkår.filtereVilkårMedResultat(behandlingId,
                                          lagredeVilkårsvurderinger,
                                          vilkårstyper,
                                          Vilkårsresultat.IKKE_TATT_STILLING_TIL)
 
-        if (erAlleVilkårVurdert(behandling, lagredeVilkårsvurderinger, vilkårstyper)) {
+        if (OppdaterVilkår.erAlleVilkårVurdert(behandling, lagredeVilkårsvurderinger, vilkårstyper)) {
             stegService.håndterVilkår(behandling).id
         } else if (vilkårUtenVurdering.isNotEmpty() && behandling.steg != StegType.VILKÅR) {
             stegService.resetSteg(behandling.id, StegType.VILKÅR)
@@ -100,13 +101,13 @@ class VurderingService(private val behandlingService: BehandlingService,
     private fun nullstillVilkårMedNyeHovedregler(behandlingId: UUID,
                                                  vilkårsvurdering: Vilkårsvurdering): VilkårsvurderingDto {
         val metadata = hentHovedregelMetadata(behandlingId)
-        val nyeDelvilkår = lagNyeDelvilkår(hentVilkårsregel(vilkårsvurdering.type), metadata)
+        val nyeDelvilkår = hentVilkårsregel(vilkårsvurdering.type).lagNyeDelvilkår(metadata)
         val delvilkårsvurdering = DelvilkårsvurderingWrapper(nyeDelvilkår)
         return vilkårsvurderingRepository.update(vilkårsvurdering.copy(resultat = Vilkårsresultat.IKKE_TATT_STILLING_TIL,
                                                                        delvilkårsvurdering = delvilkårsvurdering)).tilDto()
     }
 
-    private fun oppdaterVilkårsvurderingTilIkkeVurder(behandlingId: UUID,
+    private fun oppdaterVilkårsvurderingTilSkalIkkeVurderes(behandlingId: UUID,
                                                       vilkårsvurdering: Vilkårsvurdering): VilkårsvurderingDto {
         val metadata = hentHovedregelMetadata(behandlingId)
         val nyeDelvilkår = hentVilkårsregel(vilkårsvurdering.type).initereDelvilkårsvurderingMedVilkårsresultat(metadata,
@@ -168,110 +169,8 @@ class VurderingService(private val behandlingService: BehandlingService,
         return vilkårsvurderingRepository.insertAll(nyeVilkårsvurderinger)
     }
 
-    fun opprettNyeVilkårsvurderinger(behandlingId: UUID,
-                                     metadata: HovedregelMetadata): List<Vilkårsvurdering> {
-        return alleVilkårsregler
-                .flatMap { vilkårsregel ->
-                    if (vilkårsregel.vilkårType == VilkårType.ALENEOMSORG) {
-                        metadata.søknad.barn.map {
-                            lagNyVilkårsvurdering(vilkårsregel, metadata, behandlingId, it.id)
-                        }
-                    } else {
-                        listOf(lagNyVilkårsvurdering(vilkårsregel, metadata, behandlingId))
-                    }
-                }
-    }
-
-    private fun lagNyVilkårsvurdering(vilkårsregel: Vilkårsregel,
-                                      metadata: HovedregelMetadata,
-                                      behandlingId: UUID,
-                                      barnId: UUID? = null): Vilkårsvurdering {
-        val delvilkårsvurdering = lagNyeDelvilkår(vilkårsregel, metadata)
-        return Vilkårsvurdering(behandlingId = behandlingId,
-                                type = vilkårsregel.vilkårType,
-                                barnId = barnId,
-                                delvilkårsvurdering = DelvilkårsvurderingWrapper(delvilkårsvurdering))
-    }
-
-    private fun lagNyeDelvilkår(vilkårsregel: Vilkårsregel, metadata: HovedregelMetadata): List<Delvilkårsvurdering> {
-        return vilkårsregel.initereDelvilkårsvurderingMedVilkårsresultat(metadata,
-                                                                         resultat = Vilkårsresultat.IKKE_TATT_STILLING_TIL)
-    }
-
-    /* erAlleVilkårVurdert må vi sjekke;
-        Noen som har IKKE_TATT_STILLING_TIL -> false
-        Noen som har SKAL_IKKE_VURDERES og resten OPPFYLT -> false
-        Ellers -> true?
-    */
-
-  fun erAlleVilkårVurdert(behandling: Behandling,
-                            lagredeVilkårsvurderinger: List<Vilkårsvurdering>,
-                            vilkårstyper: List<VilkårType>): Boolean {
-        if (behandling.steg == StegType.VILKÅR) {
-            return when {
-                herNoenVurderingerIkkeTattStillingTil(behandling, lagredeVilkårsvurderinger, vilkårstyper) -> false
-                harMiksAvSkalIkkeVurderesOgOppfylt(behandling, lagredeVilkårsvurderinger, vilkårstyper) -> false
-                else -> true
-            }
-        }
-        return false
-    }
-                            lagredeVilkårsvurderinger: List<Vilkårsvurdering>,
-                            vilkårstyper: List<VilkårType>): Boolean {
-
-        if (behandling.steg == StegType.VILKÅR) {
-            val harNoenVurderingIkkeTattStillingTil = filtereVilkårMedResultat(behandling.id,
-                                                                               lagredeVilkårsvurderinger,
-                                                                               vilkårstyper,
-                                                                               Vilkårsresultat.IKKE_TATT_STILLING_TIL).isNotEmpty()
-            val harNoenVurderingSkallIkkeVurderes =
-                    filtereVilkårMedResultat(behandling.id,
-                                             lagredeVilkårsvurderinger,
-                                             vilkårstyper,
-                                             Vilkårsresultat.SKAL_IKKE_VURDERES).isNotEmpty()
-
-            return when {
-                harNoenVurderingIkkeTattStillingTil -> false
-                harNoenVurderingSkallIkkeVurderes -> !harNoenSomSkalIkkeVurderesOgRestenErOppfylt(behandling.id,
-                                                                                                  lagredeVilkårsvurderinger)
-                else -> true
-            }
-        }
-        return false
-    }
-
-
-    /* Hantering av bakåtkompatibilitet.*/
-    private fun erAktuelltVilkårType(vilkårsvurdering: Vilkårsvurdering) = VilkårType.hentVilkår().contains(vilkårsvurdering.type)
-
-
     private fun behandlingErLåstForVidereRedigering(behandlingId: UUID) =
             behandlingService.hentBehandling(behandlingId).status.behandlingErLåstForVidereRedigering()
-
-    fun filtereVilkårMedResultat(behandlingId: UUID,
-                                 lagredeVilkårsvurderinger: List<Vilkårsvurdering>,
-                                 vilkårstyper: List<VilkårType>,
-                                 resultat: Vilkårsresultat): List<Vilkårsvurdering> {
-        val aktuelleVilkår = lagredeVilkårsvurderinger
-                .filter { erAktuelltVilkårType(it) }
-
-        val antallVilkårstyper = aktuelleVilkår.map { v -> v.type }.distinct().size
-
-        require(antallVilkårstyper == vilkårstyper.size)
-        { "Forventer att det er like mange vilkår som finnes definert" }
-
-        return aktuelleVilkår.filter { it.resultat == resultat }
-
-    }
-
-    fun harNoenSomSkalIkkeVurderesOgRestenErOppfylt(behandlingId: UUID,
-                                                    lagredeVilkårsvurderinger: List<Vilkårsvurdering>): Boolean {
-
-        return lagredeVilkårsvurderinger
-                .filter { erAktuelltVilkårType(it) }
-                .filter { it.resultat != Vilkårsresultat.SKAL_IKKE_VURDERES }
-                .all { it.resultat == Vilkårsresultat.OPPFYLT }
-    }
 
 
 }
